Refactor Pyomo Dispatch
